requirements.txt
streamlit
pandas
plotly
google-cloud-firestore
openpyxl
